+++ src/Domain/Entities/Download/Base/DownloadTaskFileBase.cs
@@ -6,13 +6,13 @@ public abstract class DownloadTaskFileBase : DownloadTaskBase, IDownloadTaskProg
     /// Gets or sets the identifier of the "Metadata.Media.Id" of the media being downloaded.
     /// <remarks> This is coming from the Plex API and is used to identify the specific media part/file that is being downloaded.</remarks>
     /// </summary>
-    public required long PlexApiMediaId { get; init; }
+    public required int PlexApiMediaId { get; init; }
 
     /// <summary>
     /// Gets or sets the identifier of the "Metadata.Media.Part.Id" of the media being downloaded.
     /// <remarks> This is coming from the Plex API and is used to identify the specific media part/file that is being downloaded.</remarks>
     /// </summary>
-    public required long PlexApiPartId { get; init; }
+    public required int PlexApiPartId { get; init; }
 
     [Column(Order = 11)]
     public required string FileName { get; set; }
